requirements.txt
alembic==1.8.0
aniso8601==9.0.1
asttokens 
autopep8==1.6.0
Babel==2.10.3
backcall 
backports.functools-lru-cache 
cachelib==0.9.0
certifi 
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama 
cryptography==37.0.4
cycler==0.11.0
debugpy
decorator 
distlib==0.3.6
dnspython==2.2.1
email-validator==1.2.1
entrypoints 
executing 
filelock==3.8.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.34.4
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
infinity==1.5
intervals==0.9.2
ipykernel 
ipython 
itsdangerous==2.1.2
jedi 
Jinja2==3.1.2
jupyter-client
jupyter_core
jwt==1.3.1
kiwisolver==1.4.3
Mako==1.2.1
marko==1.2.1
markup==0.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline
nest-asyncio 
numpy==1.23.1
packaging
pandas==1.4.3
parso 
phonenumbers==8.12.52
pickleshare 
Pillow==9.2.0
platformdirs==2.5.4
prompt-toolkit 
psutil 
psycopg2==2.9.3
pure-eval 
pycodestyle==2.8.0
pycparser==2.21
Pygments
PyJWT==2.4.0
pymongo==4.3.3
pyparsing
python-dateutil 
python-dotenv==0.20.0

requests==2.28.1
scipy==1.8.1
seaborn==0.11.2
six 
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
stack-data
toml==0.10.2
tornado 
traitlets 
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.0
virtualenv-clone==0.5.7
waitress==2.1.2
wcwidth 
Werkzeug==2.1.2
wincertstore==0.2
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
